[Refactor] Divide ResultModa code as Component from lotto.js to ResultModal.js
